Fix overlay application and tpa6130a2 loading
Our overlays use #address-cells = <2> in the base_fpga_region
because we are doing address translation for the heavyweight and
lightweight bridges the same way that sopc2dts does; it might be
worth changing our overlays to use #address-cells = <1> to be more
consistent with the base socfpga dts.

Our TPA6130a2 module does not load properly when under the i2c node,
so it was placed back in the root of the device tree. It really should
belong under the i2c node, however.